pyproject.toml
[tool.poetry]
name = "aplicacao de chamada"
version = "1.0.0"
description = ""
authors = ["cadevu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.3"
psycopg2-binary = "^2.9.6"
flask-cli = "^0.4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
